Split Script Engine into it's own files & Already compiled files are skipped when starting the engine
