+++ imgui.ini
@@ -58,8 +58,8 @@ Collapsed=0
 DockId=0x0000000D,0
 
 [Window][Viewport]
-Pos=300,53
-Size=386,421
+Pos=277,53
+Size=409,421
 Collapsed=0
 DockId=0x0000000D,0
 
@@ -71,7 +71,7 @@ DockId=0x0000000E,0
 
 [Window][Hierarchy]
 Pos=0,53
-Size=298,421
+Size=275,421
 Collapsed=0
 DockId=0x0000000B,0
 
@@ -119,9 +119,9 @@ DockSpace           ID=0x04726DD8 Window=0x004E1B88 Pos=0,20 Size=1280,699 Split
   DockNode          ID=0x00000008 Parent=0x04726DD8 SizeRef=1280,670 Split=Y
     DockNode        ID=0x00000001 Parent=0x00000008 SizeRef=1280,421 Split=X Selected=0x08AA43A5
       DockNode      ID=0x00000009 Parent=0x00000001 SizeRef=1512,427 Split=X
-        DockNode    ID=0x0000000B Parent=0x00000009 SizeRef=298,427 Selected=0x29EABFBD
-        DockNode    ID=0x0000000C Parent=0x00000009 SizeRef=580,427 Split=X
-          DockNode  ID=0x0000000D Parent=0x0000000C SizeRef=386,427 CentralNode=1 Selected=0x13926F0B
+        DockNode    ID=0x0000000B Parent=0x00000009 SizeRef=275,427 Selected=0x29EABFBD
+        DockNode    ID=0x0000000C Parent=0x00000009 SizeRef=603,427 Split=X
+          DockNode  ID=0x0000000D Parent=0x0000000C SizeRef=409,427 CentralNode=1 Selected=0x13926F0B
           DockNode  ID=0x0000000E Parent=0x0000000C SizeRef=192,427 Selected=0x9B9D255D
       DockNode      ID=0x0000000A Parent=0x00000001 SizeRef=398,427 Selected=0xE7039252
     DockNode        ID=0x00000002 Parent=0x00000008 SizeRef=1280,243 Split=X Selected=0xFF2336B5
